Fix exactness problems with larger texts.
Looks like we were able to catch more problems when it comes to really
small numbers creeping in and messing up how it counts things when we
begin working with larger and larger bodies of text.

While trying to generate strings of text from an image created out of
the 1987 Constitution of the Philippines, a much larger text than the
random AI article I was initially using, may particles like "of",
"the", etc, kept appearing beside nouns that only showed up once in
the entire text. This results in weights, where, if we normalize it by
dividing each of the weights by the sum of all the weights in a given
row, we end up with numbers that need to be expressed with scientific
notation.

Of course, when you naively take the first two decimal places, see
`first-two-decimal-places`, you get rid of the scientific notation
denoting the exponent at the end. That turns it into a number greater
than 1, usually with a few decimal numbers too. When you apply that to
`inexact->exact` you get fractions with large numerators and
denominators, and when that happens while trying to generate a string,
it would throw an error.